Refactor mt-bench eval helpers into shared runtime module
